Removed SITES and TAGGIT. Moved Tags and Categories
