+++ api/blog/admin.py
@@ -1,16 +1,21 @@
 from django.contrib import admin
 
-from .models import Post
+from .models import Post, Category
+
+@admin.register(Category)
+class CategoryAdmin(admin.ModelAdmin):
+    list_display = ("name",)
 
 
 class PostAdmin(admin.ModelAdmin):
     prepopulated_fields = {"slug": ("title",)}
-    list_display = ("title", "slug", "created", "published", "is_visible")
+    list_display = ("title", "category", "slug", "created", "published", "is_visible")
     fieldsets = [
         (
             None,
             {
                 "fields": (
+                    "category",
                     "author",
                     "title",
                     "slug",
