+++ api/blog/models.py
@@ -3,8 +3,26 @@ from django.db import models
 from taggit.managers import TaggableManager
 from ckeditor_uploader.fields import RichTextUploadingField
 
+class Category(models.Model):
+    name = models.CharField(
+        blank=False, null=False, max_length=150, unique=True
+    )
+
+    def __str__(self):
+        return "{}".format(self.name)
+
+    class Meta:
+        verbose_name_plural = "categories"
+
 
 class Post(models.Model):
+    category = models.ForeignKey(
+        Category,
+        blank=True,
+        null=True,
+        on_delete=models.SET_NULL,
+        help_text="Blog category",
+    )
     author = models.ForeignKey("auth.User", on_delete=models.PROTECT)
     title = models.CharField(max_length=200)
     slug = models.SlugField(unique=True)
