+++ api/pages/admin.py
@@ -1,11 +1,6 @@
 from django.contrib import admin
 
-from .models import Page, Category, Attachment
-
-
-@admin.register(Category)
-class CategoryAdmin(admin.ModelAdmin):
-    list_display = ("name",)
+from .models import Page, Attachment
 
 
 class DocumentInline(admin.TabularInline):
@@ -14,7 +9,7 @@ class DocumentInline(admin.TabularInline):
 
 @admin.register(Page)
 class PageAdmin(admin.ModelAdmin):
-    list_display = ("title", "category", "updated_on", "is_published")
+    list_display = ("title", "updated_on", "is_published")
     list_filter = ("is_published",)
     # prepopulated_fields = {"slug": ("title",)}
     inlines = [DocumentInline]
