+++ api/requirements.txt
@@ -14,6 +14,7 @@ dj-database-url==0.5.0
 django-ckeditor==5.9.0
 django-configurations==2.2
 django-import-export==2.0.1
+django-modeladmin-reorder==0.3.1
 django-js-asset==1.2.2    # via django-ckeditor
 django-storages==1.9.1
 django-taggit==1.2.0
