+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
     volumes:
       - pgdata:/var/lib/postgresql/data
     ports:
-      - "5432:5432"
+      - "5431:5431"
     environment:
       - POSTGRES_HOST_AUTH_METHOD=trust
